chore: bump version to 2.1.316, update changelog with comprehensive feature additions (resource budgets, memory pressure, performance testing, product strategy, task funnels, rollout channels, update observability, latency masking, offline controls, shell integration, release gates, interruption policy, accessibility, versioned data model, conflict resolution, reconciliation, enterprise policy, reliability contracts, hotkey conflict detection, quarterly value map, feature tiering, TCO model, PRD

+++ src/Redball.UI.WPF/Services/WindowsCredentialSecretProvider.cs
@@ -52,7 +52,7 @@ public sealed class WindowsCredentialSecretProvider : ISecretProvider, IDisposab
                     TargetName = Marshal.StringToCoTaskMemUni(targetName),
                     CredentialBlob = Marshal.AllocCoTaskMem(byteArray.Length),
                     CredentialBlobSize = byteArray.Length,
-                    Persist = CRED_PERSIST.LOCAL_MACHINE,
+                    Persist = (uint)CRED_PERSIST.LOCAL_MACHINE,
                     UserName = Marshal.StringToCoTaskMemUni(Environment.UserName)
                 };
 
@@ -119,7 +119,7 @@ public sealed class WindowsCredentialSecretProvider : ISecretProvider, IDisposab
                     }
 
                     var byteArray = new byte[credential.CredentialBlobSize];
-                    Marshal.Copy(credential.CredentialBlob, byteArray, 0, credential.CredentialBlobSize);
+                    Marshal.Copy(credential.CredentialBlob, byteArray, 0, (int)credential.CredentialBlobSize);
                     var value = Encoding.UTF8.GetString(byteArray);
 
                     Logger.Debug("WindowsCredentialSecretProvider", $"Retrieved secret for key: {key}");
